added backend validation

+++ backend/app.js
@@ -25,85 +25,106 @@ server.use(express.text({ type: "*/*" }));
 // Add static middlewear to return files with static content
 server.use(express.static('dist'))
 
-//Handles any type of request to the exposed endpoint, sends request data to request table
+//Handles any type of request to the exposed endpoint, sends request data to request table (webhooks use this endpoint)
 server.all("/:endpoint", async (req, res) => {
   let method = req.method;
   let headers = JSON.stringify(req.headers);
   let body = req.body; //Stored in Mongo
   let endpoint = req.params.endpoint;
+  let errorMessage = '';
 
-  //Add the body to Mongo and get a document ID
-  let documentId = await mongoInsert(body);
-
-  // Try adding the request to the SQL database if it fails, send 404 error
   try {
+    if (!await pgApi.basketExists(endpoint)) {
+      errorMessage = "Endpoint does not exist."
+      throw new Error(errorMessage);
+    }
+    
+    //Add the body to Mongo and get a document ID
+    let documentId = await mongoInsert(body);
+
+    // Try adding the request to the SQL database if it fails, send 404 error
     let requestAdded = await pgApi.addRequest(
       endpoint,
       method,
       headers,
       documentId
     );
-    if (!requestAdded) throw new Error("Request couldn't be added.");
+    if (!requestAdded) {
+      errorMessage = "Request couldn't be added."
+      throw new Error(errorMessage);
+    }
 
     res.status(204).send();
   } catch (e) {
     console.error(e);
-    res.status(404).send();
+    res.status(404).send(errorMessage);
   }
 });
 
 //Handles requests to clear the basket
 server.put("/api/baskets/:endpoint", async (req, res) => {
-  //Don't allow non-local requests to this endpoint
-  if (!req.headers.host.includes(HOST)) {
-    res.status(403).send("API access denied");
-  }
-
   let endpoint = req.params.endpoint;
+  let errorMessage = '';
 
   try {
+    if (!await pgApi.basketExists(endpoint)) {
+      errorMessage = "Endpoint does not exist."
+      throw new Error(errorMessage);
+    }
+
     let basketCleared = await pgApi.clearBasket(endpoint);
-    if (!basketCleared) throw new Error("Basket couldn't be cleared.");
+    if (!basketCleared) {
+      let errorMessage = "Basket couldn't be cleared.";
+      throw new Error(errorMessage);
+    }
 
     res.status(204).send();
   } catch (e) {
     console.error(e);
-    res.status(404).send();
+    res.status(404).send(errorMessage);
   }
 });
 
 // Handles requests to delete a basket
 server.delete("/api/baskets/:endpoint", async (req, res) => {
-  //Don't allow non-local requests to this endpoint
-  if (!req.headers.host.includes(HOST)) {
-    res.status(403).send("API access denied");
-  }
-
   let endpoint = req.params.endpoint;
+  let errorMessage = '';
 
   try {
+    if (!await pgApi.basketExists(endpoint)) {
+      errorMessage = "Endpoint does not exist."
+      throw new Error(errorMessage);
+    }
+
     let basketDeleted = await pgApi.deleteBasket(endpoint);
-    if (!basketDeleted) throw new Error("Basket couldn't be deleted.");
+    if (!basketDeleted) {
+      errorMessage = "Basket couldn't be deleted.";
+      throw new Error(errorMessage);
+    }
 
     res.status(204).send();
   } catch (e) {
     console.error(e);
-    res.status(404).send();
+    res.status(404).send(errorMessage);
   }
 });
 
 // Handles requests to get all of the requests in a basket
 server.get("/api/baskets/:endpoint", async (req, res) => {
-  //Don't allow non-local requests to this endpoint
-  if (!req.headers.host.includes(HOST)) {
-    res.status(403).send("API access denied");
-  }
-
   let endpoint = req.params.endpoint;
+  let errorMessage = '';
 
   try {
+    if (!await pgApi.basketExists(endpoint)) {
+      errorMessage = "Endpoint does not exist."
+      throw new Error(errorMessage);
+    }
+    
     let requests = await pgApi.getRequests(endpoint);
-    if (!requests) throw new Error("Requests couldn't be fetched.");
+    if (!requests) {
+      errorMessage = "Requests couldn't be fetched.";
+      throw new Error(errorMessage);
+    }
 
     for (let i = 0; i < requests.length; i++) {
       if (requests[i].id) {
@@ -119,21 +140,17 @@ server.get("/api/baskets/:endpoint", async (req, res) => {
       .send(JSON.stringify(requests));
   } catch (e) {
     console.error(e);
-    res.status(404).send();
+    res.status(404).send(errorMessage);
   }
 });
 
 // Handles requests to create a new basket
 server.post("/api/baskets/:endpoint", async (req, res) => {
-  //Don't allow non-local requests to this endpoint
-  if (!req.headers.host.includes(HOST)) {
-    res.status(403).send("API access denied");
-  }
-
   let endpoint = req.params.endpoint;
+  let errorMessage = '';
 
   try {
-    if (await pgApi.isDuplicateBasket(endpoint)) {
+    if (await pgApi.basketExists(endpoint)) {
       // 403 CONFLICT
       res.status(403).send("Could not create basket: endpoint already exists.");
     }
@@ -154,30 +171,32 @@ server.post("/api/baskets/:endpoint", async (req, res) => {
     }
 
     let newBasket = await pgApi.createBasket(endpoint);
-    if (!newBasket) throw new Error("Couldn't create basket.");
+    if (!newBasket) {
+      errorMessage = "Couldn't create basket.";
+      throw new Error(errorMessage);
+    }
 
     res.status(201).send();
   } catch (e) {
     console.error(e);
-    res.status(404).send();
+    res.status(404).send(errorMessage);
   }
 });
 
 // Handles requests to create a new url endpoint
 server.get("/api/new_url_endpoint", async (req, res) => {
-  //Don't allow non-local requests to this endpoint
-  if (!req.headers.host.includes(HOST)) {
-    res.status(403).send("API access denied");
-  }
-
+  let errorMessage = '';
   try {
     let newURLEndpoint = await pgApi.getNewURLEndpoint();
-    if (!newURLEndpoint) throw new Error("Couldn't generate new url endpoint.");
+    if (!newURLEndpoint) {
+      errorMessage = "Couldn't generate new url endpoint."
+      throw new Error(errorMessage);
+    }
 
     res.json(newURLEndpoint);
   } catch (e) {
     console.error(e);
-    res.status(404).send();
+    res.status(404).send(errorMessage);
   }
 });
